backend/requirements.txt
fastapi[standard]
python-dotenv
google.generativeai
pytest
httpx